# Use vectorize instead of jit in EOS