This PR replaces the use of the `@jit` numba decorator with the `@vectorize` decorator in the eos calculation. According to the [numba docs](https://numba.pydata.org/numba-doc/latest/user/vectorize.html), `@vectorize` is the recommended way to create a numpy ufunc, which I believe is what is wanted here.

The speedup on my laptop is about 3x:
```python
import numpy as np
from pop_tools import eos

N = 10_000_000
salt = 30 + 10 * np.random.random_sample(N)
theta = -2 + 30 * np.random.random_sample(N)
pres = 5000 * np.random.random_sample(N)

%timeit eos(salt, theta, pressure=pres)
# master -> 326 ms ± 2.65 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# this PR -> 92.2 ms ± 855 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
I also removed what seems to be an unnecessary if clause regarding `xarray.apply_ufunc`.